✨ feat: 更新 practice02 类以增强目录删除功能
- 修改 deleteDirectory 方法，增加对目录的判断，确保能够正确删除目录及其内容
- 优化代码逻辑，提高文件操作的准确性和可靠性

+++ src/main/java/com/myjavastudy/g_File/Exer03/practice02.java
@@ -22,7 +22,7 @@ public class practice02 {
             return;
         }else if(file.isFile()) {
             file.delete();
-        } else {
+        } else if(file.isDirectory()){
             File[] fileList = file.listFiles();
             for (File f : fileList) {
                 deleteDirectory(f);
